feat(remix-dev/cli/init): add `--no-delete` flag (#3573)

+++ packages/remix-dev/cli/commands.ts
@@ -56,7 +56,13 @@ export async function create({
   spinner.clear();
 }
 
-export async function init(projectDir: string) {
+type InitFlags = {
+  deleteScript?: boolean;
+};
+export async function init(
+  projectDir: string,
+  { deleteScript = true }: InitFlags = {}
+) {
   let initScriptDir = path.join(projectDir, "remix.init");
   let initScript = path.resolve(initScriptDir, "index.js");
 
@@ -79,7 +85,9 @@ export async function init(projectDir: string) {
   try {
     await initFn({ isTypeScript, packageManager, rootDirectory: projectDir });
 
-    await fse.remove(initScriptDir);
+    if (deleteScript) {
+      await fse.remove(initScriptDir);
+    }
   } catch (error) {
     if (error instanceof Error) {
       error.message = `${colors.error("🚨 Oops, remix.init failed")}\n\n${
